Handle git sha tarballs
This allows us to have urls like this in the requirements file:

https://github.com/jsocol/commonware/archive/b5544185b2d24adc1eb512735990752400ce9cbd.zip#egg=commonware

which is a tarball from github for a specific commit of a project.

If this is in the requirements file, the peep will download it every
time because there's no way to know what version it is plus versions
don't really have a lot of meaning for code being pulled from a
repository (many commits all have the same "version number").