Drastically refactor for readability.
peep_install() had grown far too much local state, and even I couldn't figure out what was going on anymore. By turning all the data structures inside out and encapsulating everything within each requirement (rather than having a bunch of hashes and sets off to the side), we reduce the number of locals in peep_install(), reduce the number of `for` loops down to one (rather than repeating them in each utility function), and surface the patterns that have emerged, such as various kinds of failure printing a header, a series of errors, and a footer.

* Temporarily break tests.
* Bump version to 2.0, both for the significance of the code change and for the upcoming backward-incompatible changes.
* Temporarily (or forever?) ignore command-line args during package downloading.
* Catch a conceivable race condition in which additional files get added to the temp dir. Probably wouldn't happen since it's not linked anywhere, but hey, it's security software.