Give each requirement its own temp dir.
This closes a potential security hole in which 2 requirements have the same filename, one matches its hash and the other doesn't, but the non-matching one got created later and thus gets installed. This was not an issue in pre-2.0 versions of peep in which all files were downloaded, and then all hashes were checked.