Keep pip 0.7 from freaking out when installing the test package.
It said "Directory 'useless' is not installable. File 'setup.py' not found." To fix that, we change our HTTP server to not have to be in the tests/packages/ dir to serve files from there. Then pip doesn't get confused.

Also introduce a manual way to run the test server, handy for debugging.