Travis: Check for flake8 warnings before running the test suite
If flake8 finds issues, the Travis run will fail early, and the tox
tests will not be run. flake8, like pep8 has a default max line length
of 80 characters, which is annoying, so this has been increased to 140
characters. Ideally we'd just disable the line length check entirely,
but by specifying an 'ignore' option in tox.ini, it overrides the
default ignore list, so we would have to keep it up to date, which would
be a pain. We can always raise the threshold to 999 characters, were we
to get undesired line length warnings in the future.