Handle pip InstallationError exceptions more gracefully
`InstallationError` covers various "not peep's fault" errors, including
no valid distribution being found for the requested package version,
the requirements file being missing, or the contents not being valid:
https://github.com/pypa/pip/blob/8.0.2/pip/exceptions.py
https://github.com/pypa/pip/search?q="raise+InstallationError"&type=Code

Previously these would have resulted in the "peep had a problem" output,
which requests that users file an issue - when these cases are not
peep's fault. Now the original pip error is output as-is, without the
peep traceback.

The tests rely on the exit code for unhandled exceptions being different
from that for handled exceptions (fixed in the previous commit).

Fixes #117.