Travis: Run the tests for each Python version in parallel
Previously the Travis tests were run in a single ~33 minute job.
With this change the wall time for the test run is ~9 minutes, and the
cumulative time is still only ~33-34 minutes.

The TOXENV variable is set using a bit of a hack, since tox does not
support wildcards (either with TOXENV or when using `-e`). The output
of `--listenvs` is filtered by the current Python version (in the form
`pyNN`) and then converted from multiple lines to a comma-separated
list.